move some functions to utilities
